refactor: split slugger constants
