+++ src/constants.mjs
@@ -11,16 +11,6 @@ export const DOC_NODE_CHANGELOG_URL =
 export const DOC_API_STABILITY_SECTION_REF_URL =
   'documentation.html#stability-index';
 
-// These are string replacements specific to Node.js API docs for anchor IDs
-export const DOC_API_SLUGS_REPLACEMENTS = [
-  { from: /node.js/i, to: 'nodejs' }, // Replace Node.js
-  { from: /&/, to: '-and-' }, // Replace &
-  { from: /[/_,:;\\ ]/g, to: '-' }, // Replace /_,:;\. and whitespace
-  { from: /--+/g, to: '-' }, // Replace multiple hyphens with single
-  { from: /^-/, to: '' }, // Remove any leading hyphen
-  { from: /-$/, to: '' }, // Remove any trailing hyphen
-];
-
 // https://github.com/nodejs/node/blob/main/doc/api/cli.md#options
 // This slug should reference the section where the available
 // options are defined.
